requirements.txt
casadi==3.6.5
decorator==5.1.1
dubins @ git+https://github.com/rdesc/pydubins.git@master
fonttools==4.51.0
llvmlite==0.42.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
mpld3==0.5.9
networkx==3.1
numba==0.59.1
numpy==1.26.4
opencv-python==4.9.0.80
packcircles==0.14
pandas==1.5.3
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
ptyprocess==0.7.0
pymunk==6.7.0
pyparsing==3.0.9
pytest==7.4.0
PyYAML==6.0.1
SciencePlots==2.1.1
scikit-image==0.22.0
scipy==1.12.0
seaborn==0.12.2
setuptools==68.2.2
shapely==2.0.1
six==1.16.0
sknw==0.15
stack-data==0.2.0
tornado==6.3.3
tqdm==4.66.2
gdown